Changed Vec2 from float to int

+++ dependencies/MathLib/Vector2.h
@@ -6,11 +6,11 @@ using namespace std;
 class Vector2
 {
 public:
-	float mX;
-	float mY;
+	int mX;
+	int mY;
 
 	Vector2();
-	Vector2(float x, float y);
+	Vector2(int x, int y);
 	Vector2 operator + (Vector2 & other);
 	Vector2 operator - (Vector2 & other);
 	Vector2 operator * (float other);
